Struct stc_ts_type_ops::union_normalization::ObjectUnionNormalizer

pub struct ObjectUnionNormalizer {
    pub preserve_specified: bool,
}

See https://github.com/dudykr/stc/blob/e8f1daf0e336d978a1de5479ad9676093faf5921/crates/stc_ts_type_checker/tests/conformance/expressions/objectLiterals/objectLiteralNormalization.ts

Fields
preserve_specified: bool
Implementations

impl ObjectUnionNormalizer

fn find_keys(&self, types: &[Type]) -> IndexSet<Vec<JsWord>>

We need to know shape of normalized type literal.

We use indexset to remove duplicate while preserving order.

fn find_keys_of_type(&self, ty: &Type) -> IndexSet<Vec<JsWord>>

fn normalize_fn_types(&mut self, ty: &mut Type)

fn normalize_call_signatures(&self, ty: &mut Type)

TODO(kdy1): Add type parameters.

fn normalize_keys(&self, types: &mut Vec<Type>)

types: Types of a union.